requirements.txt
fastapi
uvicorn[standard]
gunicorn
httpx
authlib
python-jose[cryptography]
pydantic-settings
itsdangerous
cachetools
